Create Makefiles for missing benchmarks and additional SYCL fixes
Session 3: Makefile Creation (+3 working benchmarks)
Added top-level Makefiles for 8 benchmarks that were missing them:
- logic-resim-sycl: ✓ Compiles successfully (Simulation/main)
- mf-sgd-sycl: ✓ Compiles successfully (src/main)
- miniFE-sycl: ✓ Compiles successfully (src/miniFE.x)
- daphne-sycl: Partial (needs CUDA headers)
- dwconv1d-sycl: Requires PyTorch XPU
- hpl-sycl: CMake wrapper created
- si-sycl: CMake wrapper created
- snicit-sycl: Needs CLI11 library

Compiler Standardization (Session 3):
- Updated nested Makefiles from clang++ to icpx:
  - daphne-sycl/src/points2image/Makefile
  - mf-sgd-sycl/src/Makefile
  - miniFE-sycl/src/Makefile
  - logic-resim-sycl/Simulation/Makefile

Previous Session Fixes (Sessions 1-2):
Session 1: Compiler and namespace fixes (+7 benchmarks)
- axhelm-sycl: std::fabs → fabs
- crossEntropy-sycl: std::expf → expf
- cooling-sycl: Added 'f' suffix to float literals
- distort-sycl, fresnel-sycl: Added -I../../include
- heartwall-sycl: sqrt/fabs → sycl::sqrt/sycl::fabs
- gibbs-sycl: lgamma → lgammaf, sqrt → sqrtf
- Compiler fixes: 33 benchmarks from clang++/g++ to icpx

Session 2: FP64→FP32 conversion (+1 benchmark)
- cmp-sycl: Complete double→float conversion for Arc A770 compatibility
  - Applied automated float literal conversion
  - Fixed macro constants: FACTOR, EPSILON
  - Fixed malformed literals (0.2f5f → 0.25f)

Additional benchmarks converted to FP32:
- che-sycl, complex-sycl, d3q19-bgk-sycl, dp-sycl

Makefile improvements across multiple benchmarks:
- halo-finder, hbc, hypterm, kmc, mcmd, mriQ, pso, sc, spmv
- tpacf, tqs, xlqc, xsbench

Overall Impact (All Sessions):
- Starting: 256/444 (57.7%)
- After Session 1: 375/444 (84.5%)
- After Session 2: 376/444 (84.7%)
- After Session 3: 379/444 (85.4%) - 3 new working benchmarks
- Total improvement: +123 benchmarks (+27.7%)

+++ ORG/SYCL/cmp-sycl/main.cpp
@@ -97,7 +97,7 @@ compute_semblances(sycl::nd_item<1> &item,
     for(int j=0; j < _w; j++) _num[j] = 0.0f;
 
     for(int t_id=t_id0; t_id < t_idf; t_id++) {
-      real t = sycl::sqrtf(_t0 + _c * h[t_id]) * _idt;
+      real t = sycl::sqrt(_t0 + _c * h[t_id]) * _idt;
 
       int it = (int)( t );
       int ittau = it - _tau;
